Initial drafts of Cherubic hymn in E minor
Currently don't recall composer/arranger of this one, but know that it's
in E minor. The one that is currently done in Slavonic at our parish,
the goal here is to provide an English version that can be done
alongside it instead of a completely different (Byzantine) one that we
currently do.