Progress on akathist hymn
First multilingual LaTeX document. Usually use xelatex, but started off
with standard ClassicThesis example document, and switching to xelatex
affected that badly, as well as things like the overline text bar not
being available in certain fonts I use with xelatex. That said, to get
sane input handling for Russian and Greek, probably need to get xelatex
with this document.